Absorb the comm manager into the kernel and fix future handling
