+++ src/index.ts
@@ -4,7 +4,6 @@
 
 export * from './config';
 export * from './contents';
-export * from './comm';
 export * from './ikernel';
 export * from './isession';
 export * from './kernel';
